Update to syslog 5.0
Make the crate depend on the most recent release of syslog.

We need to account for some API changes. The most significant
among them may be the absence of a method taking the severity as
an argument. Instead each severity grade now has its own method:
``crit()``, ``err()``, ``warn()`` etc.